fix: do not send token in formData or in body request and access it from the cookies, and add sepearte authorSchema file and use it in post schema, posts schema

+++ app/api/delete-post/route.js
@@ -1,14 +1,31 @@
 import axios from "axios";
+import { cookies } from "next/headers";
 
 export async function DELETE(request) {
   try {
     const body = await request.json();
 
+    const cookieStore = cookies();
+    const userDataCookie = cookieStore.get("userData");
+    if (!userDataCookie) {
+      return new Response(
+        JSON.stringify({ message: "Unauthorized - No auth token found" }),
+        {
+          status: 401,
+          headers: {
+            "Content-Type": "application/json",
+          },
+        }
+      );
+    }
+    const userData = JSON.parse(userDataCookie.value);
+    const token = userData.token;
+
     const response = await axios.delete(
       `https://tarmeezacademy.com/api/v1/posts/${body.id}`,
       {
         headers: {
-          Authorization: `Bearer ${body.token}`,
+          Authorization: `Bearer ${token}`,
           "Content-Type": "application/json",
         },
         withCredentials: true,
